Added a basic manifest file to packages
For each source artifact produced (i.e. not for things like generated doc), add a `manifest.mf` file that contains details of the project built e.g. GAV co-ordinates, build time etc. The manifest file is added to the root of the

In the future, could use the manifest to sign, make module dependencies explicit etc.