Updated gradle/overture versions supported
Now building the plugin with Gradle 4.10.3 and Overture 2.6.4

Have verified that the plugin can be used with specification projects using the following Gradle versions:
- 4.2
- 4.6
- 4.10.3
- 5.0

The use of Groovy was causing issues in compilation ordering that newer versions of Gradle did not like. Have, therefore, replaced Groovy code with Kotlin and used the kotlinx.html dsl for all html generation.