Added capability to expected pre/post/inv failures in tests
In some circumstances, it can be useful to write a test where a precondition, postcondition or invariant check is expected to fail. Without annotations in VDM there is no particularly clean way to indicate this. Such expected failures can be signalled to the test runner, by including one of the following strings in the name of the test:
- ExpectPreconditionFailure
- ExpectPostconditionFailure
- ExpectInvariantFailure