Generated launch files for failing tests
When tests fail it can be time consuming to configure Overture to run
and debug those tests. This change causes a `launch` file to be
generated in `${project.buildDir}/vdm/testLaunch`. This enables the test
to be run using a single click in Overture. By default the launch files
are generated for failing tests, but this can be configured, so that
launch files are generated for all tests or none.