Created manifest jar to reduce size of test class path
When running forked Java process to execute tests on Windows the length of
the class path is unsupported. The javaexec api does not allow wildcards,
so must create a manifest jar. The sole purpose of this jar is to store the
full class path (as URIs to their existing location).