Refactored method for passing plugin to functional tests
Functional tests were failing as Gradle's automatic test kit plugin
injection was not providing a jar to use within the forked test
execution. Refactored to old-school mechanism of passing the plugin
under test, i.e. generating a manifest jar and adding that to the
buildscript dependencies.

This approach has the added advantage that the projects generated
by the functional tests can be run from the command line, which has
the potential to facilitate debugging.